requirements.txt
ccxt==4.3.89
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
python-dotenv==1.0.1
requests==2.31.0
schedule==1.2.1
colorlog==6.8.2
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
python-multipart==0.0.9
